requirements.txt
flask==2.3.2
flask-bcrypt==1.0.1
Flask-JWT-Extended
pymongo==4.3.3
python-dotenv==1.0.0
marshmallow
"pymongo[srv]"